requirements.txt
click==8.3.0
fastapi==0.119.1
numpy
opencv_python==4.12.0.88
Pillow==12.0.0
pydantic==2.12.3
python-dotenv==1.1.1
Requests==2.32.5
shap==0.49.1
starlette==0.48.0
tabulate==0.9.0
torch==2.7.1
transformers==4.57.1
ultralytics==8.3.218
uvicorn==0.38.0
